Move request handling helper functions into utils module.
